ARM: vexpress: move timer and sched_clock init to tile code
The A15 tile only use the v2m timers and versatile sched clock
as a fallback, in the case the architected timers are not
available.

Create a timer_init() vector in the tile descriptor to delegate
the timer initialization to the tile code. The sched_clock
initialization is pushed to the tile init_early() code.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>