ARM: local timers: reserve local_timer_register() to SMP
When running an SMP_ON_UP emabled kernel on UP, we want to be able
to detect that a local timer is not going to be used (local timers
are only used on SMP platforms), so we can register it as a global
timer instead.

Return -ENXIO when is_smp() returns false.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>