KVM: Bugfix: hyp_idmap_del did not clear PGDs
The pgd_clear function was defined to a no-op, since the semantics
apparently is to use the PUD (pointing to the same bytes) for this
operation.

Interestingly enough this bug showed up as a BUG_ON() in the
hyp_create_mapping function when using a 2G/2G split for user and
kernel, since we were now all of the sudden mapping virtuall addresses
where the identity mapping had previously taken place (at 0x8.....).

I purposely leave the do { ... } while (a, b, c); strutcture in here
since the rest of the file uses that and let's keep these changes
limited to KVM right now.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>