ARM: vexpress: remove fragile UART detection mechanism
The code to 'guess' the UART base address for vexpress platforms reads
from the Configuration Base Register. This is not present on all
Versatile Express platforms and the attempt to access co-processor 15
with opcode 4 causes an abort.

This patch replaces the fuzzy guesswork code with the original, hard coded
value that is appropriate for Cortex A9. A subsequent patch is provided
in this series to choose an appropriate location using Kconfig

Signed-off-by: Jonathan Austin <jonathan.austin@arm.com>