ARM: vexpress: Add RS1 UART for decompressor, selectable via DEBUG_LL
The first UART is located at a different address on RS1 and CA9X4 vexpress
motherboards.

This patch uses the same #defines as used when enabling DEBUG_LL in order
to select which UART location to use when printing output from the
decompressor. If DEBUG_LL is not selected it prints nothing, instead of
attempting to guess a location.

Signed-off-by: Jonathan Austin <jonathan.austin@arm.com>