ARM: Fail early when using CONFIG_MMU on v7 systems without an MMU
It is possible to boot CONFIG_MMU Linux on a system with an ARMv7 processor
that doens't have an MMU. In this case we should fail early and obviously.

This patch represents a simple/clear way to achieve this. However, because
CONFIG_ARM_LPAE depends on both CONFIG_MMU and CONFIG_ARM_V7 it would be
possible to interleave this test with the LPAE one above, avoiding reading
ID_MMFR0 twice. Comments are welcome about whether it is worth sacrificing
the independence of these tests in order to reduce the instruction count.