ARM: KVM: Do not trap on WFE execution
We currently trap both WFI and WFE, and treat WFE exactly as WFI.
This is not a very nice thing to do in an SMP configuration, as
any spinlock contention on the guest will result in a massive
performance degradation.

WFE is usually used for very short periods, so staying on the
physical CPU is likely to be the best choice.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>