ARM: KVM: Module unloading support
If we don't disable the Hyp-mode MMU when unloading the KVM module other
modules relying on the Hyp-mode MMU to be disabled.

This is going to change in two ways:

First, the init id-map code is going to go away in favor of
section-based id-mapping.

Second, we are not going to use the SMC call in the future, but rather
an HVC instruction to take control of Hyp mode.  We need, however, a
method to setup the original init code again to support module
unloading.  It is useful, however, to add this support at this point
since we will (a) remember to support unloading (b) benefit from shorter
debug cycles.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>