ARM: KVM: More careful IPA lookup for ls-emulation
Some load/store instructions cannot be emulated using information in the
HSR, so the instruction must be loaded and decoded.  For that purpose we
store the IPA of the faulting instruction so we can retrieve it from the
C-emulation code.  However, (as Rusty Russell pointed out) a wide Thumb
instruction at a 16-bit aligned address on the end of a page requires
two IPAs.

Extremely rare case that we should support nonetheless.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>