ARM: KVM: adjust_itstate on guest exceptions
When an exception is taken from a guest running in Thumb-2 mode and
that instruction is emulated and skipped, the ITSTATE field stored
inside the CPSR must be updated by KVM, as the instruction is never
advanced by hardware.

This can cause bugs like ending up in a state where you incorrectly don't or do
execute instructions, or even worse, as was the case in some GIC i/o code,
where one IT-block immediately follows another one, causes the IT-block to
generate an undefined exception on the IT instruction itself.

Note: This patch does touch a tiny-bit of unrelated code, but it's going to be
flattened for next KVM release anyhow.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>