ARM: KVM: upgrade shareability to the inner domain in stage2 translation
On an SMP host, it is important that a write from a VCPU running on a
given CPU is observable from the same VCPU after being migrated to
another CPU.

Failure to do so could result in horrible things happening when
running a UP guest on a SMP host.

Set the shareability to "inner-shareable" in the stage2 translation
tables to upgrade the guest shareability.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>