kvm tools: Support cross compilation
Allow CROSS_COMPILE to be used to prefix CC as is done in the kernel
Makefile. If CROSS_COMPILE is unset it has no effect, and still allows
CC to be overridden.

We need to fix a few places to use ARCH instead of uname_M directly, so
that the overridden setting of ARCH takes effect.

Signed-off-by: Michael Ellerman <michael@ellerman.id.au>
Signed-off-by: Pekka Enberg <penberg@kernel.org>