ARM: KVM: vgic: remove locking from interrupt handler
Taking the distributor lock in the VGIC maintainance handler
has the unfortunate consequence of forcing all the other users
of that lock to disable interrupts while this lock is held.

This lock is only held to update the irq_active bit for an EOIed
level interrupt. If we remove the locking, there is a race with
the queuing of an interrupt in __kvm_sync_to_cpu(), where we check
if the interrupt is already active. Two possibilities:

- The queuing is occuring on the same vcpu: cannot happen, as we're
  already in the context of this vcpu, and executing the handler
- The interrupt has been migrated to another vcpu, and we ignore
  this interrupt for this run. Big deal. It is still pending though,
  and will get considered when this vcpu exits.

We can then safely remove that lock from vgic_maintainance_handler()
and convert the other users not to disable interrupts.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>