KVM: ARM: Remove TODO comment
Copying the guest instruction into the host is *sloow*, so while we
_may_ save a few pages with KSM here, it's going to be too expensive on
every page fault if there are may page faults with register write back,
which is more likely than the pages to be identical with something else
over a large amount of pages.

If anyone want to profile this later on, they should feel free, but it's
not on my todo list anymore.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>